+++ Release Notes.md
@@ -6,6 +6,7 @@
 ### ✨ New features
 
 * `FormText` is a new view to present text in `List` forms.
+* `FormTextEditor` is a new view to present text editors in `List` forms.
 * `FormTextField` is a new view to present text fields in `List` forms.
 * `ActionButton` is a new button that resembles the action button in e.g. Contacts.
 * `ActionButtonStyle` is a new button style that can be used for action buttons.